+++ iModelBridgeCore/Imagepp/all/gra/hrf/src/HRFAdaptStripToImage.cpp
@@ -2,15 +2,15 @@
 //:>
 //:>     $Source: all/gra/hrf/src/HRFAdaptStripToImage.cpp $
 //:>
-//:>  $Copyright: (c) 2014 Bentley Systems, Incorporated. All rights reserved. $
+//:>  $Copyright: (c) 2015 Bentley Systems, Incorporated. All rights reserved. $
 //:>
 //:>+--------------------------------------------------------------------------------------
 //-----------------------------------------------------------------------------
 // Class HRFAdaptStripToImage
 //-----------------------------------------------------------------------------
 
-#include <ImagePP/h/hstdcpp.h>
-#include <ImagePP/h/HDllSupport.h>
+#include <ImagePPInternal/hstdcpp.h>
+
 #include <Imagepp/all/h/HFCAccessMode.h>
 
 #include <Imagepp/all/h/HCDPacket.h>
@@ -128,9 +128,9 @@ HRFAdaptStripToImage::~HRFAdaptStripToImage()
 // ReadBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptStripToImage::ReadBlock(uint32_t pi_PosBlockX,
-                                        uint32_t pi_PosBlockY,
-                                        Byte* po_pData,
+HSTATUS HRFAdaptStripToImage::ReadBlock(uint64_t pi_PosBlockX,
+                                        uint64_t pi_PosBlockY,
+                                        Byte*  po_pData,
                                         HFCLockMonitor const* pi_pSisterFileLock)
     {
     HPRECONDITION (m_AccessMode.m_HasReadAccess);
@@ -203,8 +203,8 @@ HSTATUS HRFAdaptStripToImage::ReadBlock(uint32_t pi_PosBlockX,
 // ReadBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptStripToImage::ReadBlock(uint32_t            pi_PosBlockX,
-                                        uint32_t            pi_PosBlockY,
+HSTATUS HRFAdaptStripToImage::ReadBlock(uint64_t            pi_PosBlockX,
+                                        uint64_t            pi_PosBlockY,
                                         HFCPtr<HCDPacket>&  po_rpPacket,
                                         HFCLockMonitor const* pi_pSisterFileLock)
     {
@@ -216,9 +216,9 @@ HSTATUS HRFAdaptStripToImage::ReadBlock(uint32_t            pi_PosBlockX,
 // ReadBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptStripToImage::WriteBlock(uint32_t     pi_PosBlockX,
-                                         uint32_t     pi_PosBlockY,
-                                         const Byte* pi_pData,
+HSTATUS HRFAdaptStripToImage::WriteBlock(uint64_t     pi_PosBlockX,
+                                         uint64_t     pi_PosBlockY,
+                                         const Byte*  pi_pData,
                                          HFCLockMonitor const* pi_pSisterFileLock)
     {
     HPRECONDITION (m_AccessMode.m_HasWriteAccess || m_AccessMode.m_HasCreateAccess);
@@ -293,8 +293,8 @@ HSTATUS HRFAdaptStripToImage::WriteBlock(uint32_t     pi_PosBlockX,
 // ReadImage
 // Edition by Image
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptStripToImage::WriteBlock(uint32_t                 pi_PosBlockX,
-                                         uint32_t                 pi_PosBlockY,
+HSTATUS HRFAdaptStripToImage::WriteBlock(uint64_t                 pi_PosBlockX,
+                                         uint64_t                 pi_PosBlockY,
                                          const HFCPtr<HCDPacket>& pi_rpPacket,
                                          HFCLockMonitor const* pi_pSisterFileLock)
     {
@@ -308,8 +308,8 @@ HSTATUS HRFAdaptStripToImage::WriteBlock(uint32_t                 pi_PosBlockX,
 // WriteBlockRLE
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptStripToImage::WriteBlockRLE(uint32_t              pi_PosBlockX,
-                                            uint32_t              pi_PosBlockY,
+HSTATUS HRFAdaptStripToImage::WriteBlockRLE(uint64_t              pi_PosBlockX,
+                                            uint64_t              pi_PosBlockY,
                                             HFCPtr<HCDPacketRLE>& pi_rpPacketRLE,
                                             HFCLockMonitor const* pi_pSisterFileLock)
     {
@@ -364,10 +364,10 @@ HSTATUS HRFAdaptStripToImage::WriteBlockRLE(uint32_t              pi_PosBlockX,
 // ReadBlockRLE
 // Edition by Image
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptStripToImage::ReadBlockRLE  (uint32_t                 pi_PosBlockX,
-                                             uint32_t                 pi_PosBlockY,
-                                             HFCPtr<HCDPacketRLE>&    po_rpPacketRLE,
-                                             HFCLockMonitor const*    pi_pSisterFileLock)
+HSTATUS HRFAdaptStripToImage::ReadBlockRLE(uint64_t                 pi_PosBlockX,
+                                           uint64_t                 pi_PosBlockY,
+                                           HFCPtr<HCDPacketRLE>&    po_rpPacketRLE,
+                                           HFCLockMonitor const*    pi_pSisterFileLock)
     {
     HPRECONDITION(m_AccessMode.m_HasReadAccess);
     HPRECONDITION(po_rpPacketRLE->HasBufferOwnership());    // Must be owner of buffers.
