+++ iModelBridgeCore/Imagepp/all/gra/hrf/src/HRFAdaptTileToImage.cpp
@@ -2,15 +2,15 @@
 //:>
 //:>     $Source: all/gra/hrf/src/HRFAdaptTileToImage.cpp $
 //:>
-//:>  $Copyright: (c) 2014 Bentley Systems, Incorporated. All rights reserved. $
+//:>  $Copyright: (c) 2015 Bentley Systems, Incorporated. All rights reserved. $
 //:>
 //:>+--------------------------------------------------------------------------------------
 //-----------------------------------------------------------------------------
 // Class HRFAdaptTileToImage
 //-----------------------------------------------------------------------------
 
-#include <ImagePP/h/hstdcpp.h>
-#include <ImagePP/h/HDllSupport.h>
+#include <ImagePPInternal/hstdcpp.h>
+
 
 #include <Imagepp/all/h/HFCAccessMode.h>
 
@@ -116,9 +116,9 @@ HRFAdaptTileToImage::~HRFAdaptTileToImage()
 // ReadBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptTileToImage::ReadBlock(uint32_t pi_PosBlockX,
-                                       uint32_t pi_PosBlockY,
-                                       Byte* po_pData,
+HSTATUS HRFAdaptTileToImage::ReadBlock(uint64_t pi_PosBlockX,
+                                       uint64_t pi_PosBlockY,
+                                       Byte*  po_pData,
                                        HFCLockMonitor const* pi_pSisterFileLock)
     {
     HPRECONDITION (m_AccessMode.m_HasReadAccess);
@@ -188,8 +188,8 @@ HSTATUS HRFAdaptTileToImage::ReadBlock(uint32_t pi_PosBlockX,
 // ReadBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptTileToImage::ReadBlock(uint32_t            pi_PosBlockX,
-                                       uint32_t            pi_PosBlockY,
+HSTATUS HRFAdaptTileToImage::ReadBlock(uint64_t            pi_PosBlockX,
+                                       uint64_t            pi_PosBlockY,
                                        HFCPtr<HCDPacket>&  po_rpPacket,
                                        HFCLockMonitor const* pi_pSisterFileLock)
     {
@@ -201,8 +201,8 @@ HSTATUS HRFAdaptTileToImage::ReadBlock(uint32_t            pi_PosBlockX,
 // WriteBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptTileToImage::WriteBlock(uint32_t     pi_PosBlockX,
-                                        uint32_t     pi_PosBlockY,
+HSTATUS HRFAdaptTileToImage::WriteBlock(uint64_t    pi_PosBlockX,
+                                        uint64_t    pi_PosBlockY,
                                         const Byte* pi_pData,
                                         HFCLockMonitor const* pi_pSisterFileLock)
     {
@@ -277,8 +277,8 @@ HSTATUS HRFAdaptTileToImage::WriteBlock(uint32_t     pi_PosBlockX,
 // ReadBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFAdaptTileToImage::WriteBlock(uint32_t                    pi_PosBlockX,
-                                        uint32_t                   pi_PosBlockY,
+HSTATUS HRFAdaptTileToImage::WriteBlock(uint64_t                   pi_PosBlockX,
+                                        uint64_t                   pi_PosBlockY,
                                         const HFCPtr<HCDPacket>&   pi_rpPacket,
                                         HFCLockMonitor const* pi_pSisterFileLock)
     {
