+++ iModelBridgeCore/Imagepp/all/gra/hrf/src/HRFCacheRandomBlockEditor.cpp
@@ -2,15 +2,15 @@
 //:>
 //:>     $Source: all/gra/hrf/src/HRFCacheRandomBlockEditor.cpp $
 //:>
-//:>  $Copyright: (c) 2014 Bentley Systems, Incorporated. All rights reserved. $
+//:>  $Copyright: (c) 2015 Bentley Systems, Incorporated. All rights reserved. $
 //:>
 //:>+--------------------------------------------------------------------------------------
 //-----------------------------------------------------------------------------
 // Class HRFCacheRandomBlockEditor
 //-----------------------------------------------------------------------------
 
-#include <ImagePP/h/hstdcpp.h>
-#include <ImagePP/h/HDllSupport.h>
+#include <ImagePPInternal/hstdcpp.h>
+
 #include <Imagepp/all/h/HRFCacheRandomBlockEditor.h>
 #include <Imagepp/all/h/HFCMonitor.h>
 #include <Imagepp/all/h/HCDPacket.h>
@@ -58,9 +58,9 @@ HRFCacheRandomBlockEditor::~HRFCacheRandomBlockEditor()
 // ReadBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFCacheRandomBlockEditor::ReadBlock(uint32_t pi_PosBlockX,
-                                             uint32_t pi_PosBlockY,
-                                             Byte* po_pData,
+HSTATUS HRFCacheRandomBlockEditor::ReadBlock(uint64_t pi_PosBlockX,
+                                             uint64_t pi_PosBlockY,
+                                             Byte*  po_pData,
                                              HFCLockMonitor const* pi_pSisterFileLock)
     {
     HPRECONDITION (m_AccessMode.m_HasReadAccess);
@@ -123,8 +123,8 @@ HSTATUS HRFCacheRandomBlockEditor::ReadBlock(uint32_t pi_PosBlockX,
 // ReadBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFCacheRandomBlockEditor::ReadBlock(uint32_t           pi_PosBlockX,
-                                             uint32_t           pi_PosBlockY,
+HSTATUS HRFCacheRandomBlockEditor::ReadBlock(uint64_t           pi_PosBlockX,
+                                             uint64_t           pi_PosBlockY,
                                              HFCPtr<HCDPacket>& po_rpPacket,
                                              HFCLockMonitor const* pi_pSisterFileLock)
     {
@@ -201,9 +201,9 @@ HSTATUS HRFCacheRandomBlockEditor::ReadBlock(uint32_t           pi_PosBlockX,
 // WriteBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFCacheRandomBlockEditor::WriteBlock(uint32_t     pi_PosBlockX,
-                                              uint32_t     pi_PosBlockY,
-                                              const Byte* pi_pData,
+HSTATUS HRFCacheRandomBlockEditor::WriteBlock(uint64_t      pi_PosBlockX,
+                                              uint64_t      pi_PosBlockY,
+                                              const Byte*   pi_pData,
                                               HFCLockMonitor const* pi_pSisterFileLock)
     {
     HPRECONDITION (m_AccessMode.m_HasWriteAccess || m_AccessMode.m_HasCreateAccess);
@@ -228,8 +228,8 @@ HSTATUS HRFCacheRandomBlockEditor::WriteBlock(uint32_t     pi_PosBlockX,
 // WriteBlock
 // Edition by Block
 //-----------------------------------------------------------------------------
-HSTATUS HRFCacheRandomBlockEditor::WriteBlock(uint32_t                 pi_PosBlockX,
-                                              uint32_t                 pi_PosBlockY,
+HSTATUS HRFCacheRandomBlockEditor::WriteBlock(uint64_t                 pi_PosBlockX,
+                                              uint64_t                 pi_PosBlockY,
                                               const HFCPtr<HCDPacket>& pi_rpPacket,
                                               HFCLockMonitor const* pi_pSisterFileLock)
     {
